Basic display Added basic display Positions are now floats Fixed #1

+++ Logic/Position.h
@@ -6,18 +6,18 @@ public:
 	Position(){};
 	Position(int x, int y){ this->x=x; this->y = y; };
 	Position(Position &position){ x= position.x; y = position.y; };
-	int getX(){ return x; };
-	int getY(){ return y; };
-	void setX(int x) { this->x = x; };
-	void setY(int y) { this->y = y; };
-	void set(int x, int y) { setX(x); setY(y); };
+	float getX(){ return x; };
+	float getY(){ return y; };
+	void setX(float x) { this->x = x; };
+	void setY(float y) { this->y = y; };
+	void set(float x, float y) { setX(x); setY(y); };
 	//double distance(Position position){ return sqrt((double)((x - position.x) ^ 2) + (double)((y - position.y) ^ 2));};
-	double distance(Position position)
-	{ 	double disX = (position.x - x);
-		double disY = (position.y - y);
+	float distance(Position position)
+	{ 	float disX = (position.x - x);
+		float disY = (position.y - y);
 		return sqrt(pow(disX,2)+pow(disY,2));
 	};
 protected:
-	int x, y;
+	float x, y;
 };
